requirements.txt
--prefer-binary

# Flask and extensions
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Limiter==3.3.1
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Swagger-UI==4.11.1

# Database - using only pg8000 driver for Vercel compatibility
SQLAlchemy==2.0.21
pg8000==1.31.2
alembic==1.12.0


# API documentation
apispec==6.3.0
marshmallow==3.20.1

# Core functionality
PyYAML==6.0.1
python-dotenv==1.0.0
SpeechRecognition==3.10.0
deep-translator==1.11.4
gTTS==2.4.0

# Server
uvicorn==0.23.2
Werkzeug==2.2.3
gunicorn==23.0.0
firebase-admin
git-filter-repo==2.47.0
# Authentication
PyJWT==2.8.0
bcrypt==4.3.0
email_validator==2.0.0

# Storage backends for rate limiting
redis==4.3.4
limits==2.8.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pydub==0.25.1
httpcore==0.17.3
httpx==0.24.1
cryptography==41.0.4

# AI services
openai==0.28.1
google-cloud-speech==2.21.0
google-cloud-translate==3.11.1